requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
pydantic
httpx